Implement rbac

+++ services/auth-service/repository/user_repository.go
@@ -38,7 +38,7 @@ func (u *UserRepository) FindByEmail(ctx context.Context, email string) (*entity
 	}
 	return &user, nil
 }
-func (u *UserRepository) FindByID(ctx context.Context, id int) (*entity.User, error) {
+func (u *UserRepository) FindByID(ctx context.Context, id int64) (*entity.User, error) {
 	query := `SELECT id, name, email, role, created_at, updated_at FROM users WHERE id = $1`
 	var user entity.User
 	err := u.db.QueryRowContext(ctx, query, id).Scan(
